# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API ePortfolio",
      "description": "Daftar API ePortfolio",
      "version": "1.0"
    },
    "servers":[
        {
            "url": "http://127.0.0.1:8800/api",
            "description":"localhost"
        },
        {
            "url": "http://103.187.223.15:8800/api",
            "description":"dev"
        }
    ],
...
    "produces": ["application/json"],
    "paths": {
...
      "/projects":{
        "post":{
          "tags":[
            "Project"
          ],
          "summary":"Create project",
          "description":"Create a new project",
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "image":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "type":{
                      "type":"string",
                      "enum": ["academic", "non-academic"]
                    },
                    "startDate":{
                      "type":"string"
                    },
                    "endDate":{
                      "type":"string",
                      "nullable":true
                    },
                    "desc":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/projects/{id}":{
        "put":{
          "tags":[
            "Project"
          ],
          "summary":"Update Project",
          "description":"Update project",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"project id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "image":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "type":{
                      "type":"string",
                      "enum": ["academic", "non-academic"]
                    },
                    "startDate":{
                      "type":"string"
                    },
                    "endDate":{
                      "type":"string",
                      "nullable":true
                    },
                    "desc":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "delete":{
          "tags":[
            "Project"
          ],
          "summary":"Delete Project",
          "description":"Delete project",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"project id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "get":{
          "tags":[
            "Project"
          ],
          "summary":"Get Project",
          "description":"Get project data",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"project id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/projects/all/{id}":{
        "get":{
          "tags":[
            "Project"
          ],
          "summary":"Get All Projects From User",
          "description":"Get all projects data from user",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/projects/search/{keyword}":{
        "get":{
          "tags":[
            "Project"
          ],
          "summary":"Get Project by Keyword",
          "description":"Get project data by keyword",
          "parameters":[
            {
              "in":"path",
              "name": "keyword",
              "description":"keyword"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/projects/suggest/{id}":{
        "get":{
          "tags":[
            "Project"
          ],
          "summary":"Get Recent Projects Suggestion",
          "description":"Get 3 recent projects suggestion",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/projects/request/{id}":{
        "put":{
          "tags":[
            "Project"
          ],
          "summary":"Request to participate",
          "description":"Request to participate in the project",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"project id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/projects/cancelrequest/{id}":{
        "put":{
          "tags":[
            "Project"
          ],
          "summary":"Cancel Request to participate",
          "description":"Cancel request to participate in the project",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"project id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/projects/accept/{id}":{
        "put":{
          "tags":[
            "Project"
          ],
          "summary":"Accept Request to participate",
          "description":"Accept request to participate in the project",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"project id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
...
    }
 }